processing/azure_container_apps/requirements.txt
azure-core
azure-identity
azure-mgmt-appcontainers
azure-containerregistry
azure-mgmt-containerregistry
azure-mgmt-resource